--- convert-junit4-to-junit5/src/main/java/jb/convert/ast/StaticImport.java
@@ -1,14 +1,0 @@
-package jb.convert.ast;
-
-import jb.ClassName;
-
-class StaticImport {
-
-    final ClassName className;
-    final String method;
-
-    StaticImport(ClassName className, String method) {
-        this.className = className;
-        this.method = method;
-    }
-}
