+++ convert-junit4-to-junit5/src/main/java/jb/convert/ast/tools/Callback.java
@@ -1,4 +1,4 @@
-package jb.convert.ast;
+package jb.convert.ast.tools;
 
 public interface Callback {
     void call();
