+++ convert-junit4-to-junit5/src/main/java/jb/convert/ast/tools/PackageDeclarations.java
@@ -1,12 +1,12 @@
-package jb.convert.ast;
+package jb.convert.ast.tools;
 
 import com.github.javaparser.HasParentNode;
 import com.github.javaparser.ast.CompilationUnit;
 import com.github.javaparser.ast.nodeTypes.NodeWithName;
 
-class PackageDeclarations {
+public class PackageDeclarations {
 
-    static String packageNameFor(HasParentNode<?> n) {
+    public static String packageNameFor(HasParentNode<?> n) {
         return n.findAncestor(CompilationUnit.class)
                 .map(cu -> cu.getPackageDeclaration().map(NodeWithName::getNameAsString)
                         .orElse(""))
