+++ convert-junit4-to-junit5/src/test/java/jb/convert/ast/ImportDeclarationsTest.java
@@ -2,6 +2,7 @@ package jb.convert.ast;
 
 import com.github.javaparser.ast.CompilationUnit;
 import com.github.javaparser.ast.body.ClassOrInterfaceDeclaration;
+import jb.convert.ast.tools.ImportDeclarations;
 import org.junit.jupiter.api.Test;
 
 import static org.hamcrest.MatcherAssert.assertThat;
